Increase precision of julian dates
Just in case. They were floats, now they're doubles.

+++ filmulator-gui/database/dateHistogramModel.h
@@ -36,8 +36,8 @@ signals:
 
 protected:
     void emitChange() {emit dateHistoModelChanged();}
-    float m_today;
-    float m_firstDay;
+    double m_today;
+    double m_firstDay;
     std::vector<double> m_dataVector;
 };
